docs(dojo): shape WF-165 tranche D
