+++ tests/cycles/WF-130/roadmap-goalpost-policy.part04.test.ts
@@ -1,14 +1,8 @@
-import { existsSync } from 'node:fs';
-import { resolve } from 'node:path';
-import { describe, expect, it } from 'vitest';
+import { describe, it } from 'vitest';
 import {
   expectClaims,
   expectNoClaims,
   normalized,
-  read,
-  requireRecord,
-  ROOT,
-  sectionBetween,
 } from './roadmap-goalpost-policy.test-support.js';
 
 describe('WF-130 roadmap supporting contracts', () => {
@@ -46,20 +40,22 @@ describe('WF-130 roadmap supporting contracts', () => {
       'The next feature horizon remains `v8.0.0`',
       '[#477](https://github.com/flyingrobots/bijou/issues/477) has met its `112 -> 62` contract.',
       'Landed tranche A [#475](https://github.com/flyingrobots/bijou/pull/475) and landed tranche B [#478](https://github.com/flyingrobots/bijou/pull/478) each removed `25` counted violations.',
-      'Tranche C [#487](https://github.com/flyingrobots/bijou/pull/487) lowers current debt to `22` file/context and `10` code-size violations with no mock-ban or ESLint debt.',
+      'Landed tranche C [#487](https://github.com/flyingrobots/bijou/pull/487) lowers current debt to `22` file/context and `10` code-size violations with no mock-ban or ESLint debt.',
+      'Tranche D is selected to lower those ledgers to `17` and `5`.',
       'The bounded target [#468](https://github.com/flyingrobots/bijou/issues/468) landed through [#474](https://github.com/flyingrobots/bijou/pull/474)',
       'The #458 GraphQL block artifact bundle and #459 packed-cell `Surface` adapter have landed',
       '[#480](https://github.com/flyingrobots/bijou/issues/480)',
       '[WF-165](./design/WF-165-respecting-dojo-ratchet-12.md)',
       'Landed tranche B [#486](https://github.com/flyingrobots/bijou/pull/486) removed the next five smallest roots and lowered that ceiling to `42`',
-      'Tranche C [#487](https://github.com/flyingrobots/bijou/pull/487) splits the canonical example app, Notifications, Image Viewer, the PR review-status tool, and framed-app rendering',
+      'Landed tranche C [#487](https://github.com/flyingrobots/bijou/pull/487) splits the canonical example app, Notifications, Image Viewer, the PR review-status tool, and framed-app rendering',
       'lowering the live ceiling to `32`; `10` double-counted roots remain.',
+      'Tranche D now owns `Surface`, DOGFOOD i18n-debt analysis, framed-app overlays, `ui-scene-ir/1`, and the terminal differ, targeting an aggregate ceiling of `22`.',
       '[RE-036](./design/RE-036-packed-bijou-cells-surface-adapter.md)',
       '`v7.2.0` completed as a narrow stabilization and demo-integrity release',
       '`v7.2.0` milestone is complete release lineage: 0 open and 19 closed milestone items',
       '`v8.0.0` milestone is the active feature horizon: 2 open milestone items and 2 closed milestone items',
       '`v8.1.0` milestone is replay, capture, debugger, and render-witness follow-through',
-      '`v8.2.0` milestone is quality automation and Method hardening: 21 open and 2 closed milestone items',
+      '`v8.2.0` milestone is quality automation and Method hardening: 22 open and 2 closed milestone items',
       '[#485](https://github.com/flyingrobots/bijou/issues/485)',
       'No open issue is currently unmilestoned',
       'The selected `v7.2.0` DOGFOOD product pull #335 has landed',
@@ -97,53 +93,4 @@ describe('WF-130 roadmap supporting contracts', () => {
       'VISOR coordination surface: https://github.com/flyingrobots/visor',
     ]);
   });
-  it('disables Markdown line-length linting for project docs', () => {
-    const config = requireRecord(JSON.parse(read('.markdownlint.json')));
-    expect(config.MD013).toBe(false);
-    expect(config['line-length']).toBe(false);
-  });
-
-  it('publishes discoverable v7.2 release docs for the versioned package set', () => {
-    for (const path of [
-      'docs/releases/7.2.0/README.md',
-      'docs/releases/7.2.0/whats-new.md',
-      'docs/releases/7.2.0/migration-guide.md',
-    ]) {
-      expect(existsSync(resolve(ROOT, path)), path).toBe(true);
-    }
-    expectClaims(read('docs/releases/README.md'), [
-      '[Release Evidence (v7.2.0)](./7.2.0/README.md)',
-      "[What's New (v7.2.0)](./7.2.0/whats-new.md)",
-      '[Migration Guide (v7.2.0)](./7.2.0/migration-guide.md)',
-    ]);
-  });
-
-  it('links the #458 VISOR artifact bundle cycle to DX-049', () => {
-    const path = 'docs/design/DX-049-visor-artifact-bundle-proof.md';
-    const source = read(path);
-    expect(existsSync(resolve(ROOT, path))).toBe(true);
-    expectClaims(normalized('docs/ROADMAP.md'), [
-      '[`DX-049`](./design/DX-049-visor-artifact-bundle-proof.md)',
-      '`visor-artifact-bundle/1` proof',
-    ]);
-    expectClaims(normalized('docs/BEARING.md'), [
-      'Their cycle designs remain [DX-049](./design/DX-049-visor-artifact-bundle-proof.md)',
-    ]);
-    expectClaims(normalizeSource(source), [
-      'User story: [#458](https://github.com/flyingrobots/bijou/issues/458)',
-      'V8 tracker: [#457](https://github.com/flyingrobots/bijou/issues/457)',
-      '`visor-artifact-bundle/1`',
-      '`bijou-block/1`',
-      '`ui-scene-ir/1`',
-      'replay metadata',
-      'visual scene facts',
-      'Tests To Write First',
-    ]);
-    expect(
-      sectionBetween(source, '## Non-Goals', '## Bundle Contract'),
-    ).toContain('implement #459 packed-cell-to-`Surface` validation');
-  });
 });
-function normalizeSource(source: string): string {
-  return source.replace(/\s+/g, ' ').trim();
-}
